@@ -1,7 +1,9 @@
 <?php
 
-/**
+use App\Http\Controllers\Backend\DashboardController;
+
+/*
  * All route names are prefixed with 'admin.'.
  */
 Route::redirect('/', '/admin/dashboard', 301);
-Route::get('dashboard', 'DashboardController@index')->name('dashboard');
+Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard');
