@@ -1,6 +1,15 @@
 <?php
 
-/**
+use App\Http\Controllers\Frontend\Auth\LoginController;
+use App\Http\Controllers\Frontend\Auth\RegisterController;
+use App\Http\Controllers\Frontend\Auth\SocialLoginController;
+use App\Http\Controllers\Frontend\Auth\ResetPasswordController;
+use App\Http\Controllers\Frontend\Auth\ConfirmAccountController;
+use App\Http\Controllers\Frontend\Auth\ForgotPasswordController;
+use App\Http\Controllers\Frontend\Auth\UpdatePasswordController;
+use App\Http\Controllers\Frontend\Auth\PasswordExpiredController;
+
+/*
  * Frontend Access Controllers
  * All route names are prefixed with 'frontend.auth'.
  */
@@ -10,21 +19,21 @@ Route::group(['namespace' => 'Auth', 'as' => 'auth.'], function () {
     * These routes require the user to be logged in
     */
     Route::group(['middleware' => 'auth'], function () {
-        Route::get('logout', 'LoginController@logout')->name('logout');
+        Route::get('logout', [LoginController::class, 'logout'])->name('logout');
 
         //For when admin is logged in as user from backend
-        Route::get('logout-as', 'LoginController@logoutAs')->name('logout-as');
+        Route::get('logout-as', [LoginController::class, 'logoutAs'])->name('logout-as');
 
         // These routes can not be hit if the password is expired
         Route::group(['middleware' => 'password_expires'], function () {
             // Change Password Routes
-            Route::patch('password/update', 'UpdatePasswordController@update')->name('password.update');
+            Route::patch('password/update', [UpdatePasswordController::class, 'update'])->name('password.update');
         });
 
         // Password expired routes
         if (is_numeric(config('access.users.password_expires_days'))) {
-            Route::get('password/expired', 'PasswordExpiredController@expired')->name('password.expired');
-            Route::patch('password/expired', 'PasswordExpiredController@update')->name('password.expired.update');
+            Route::get('password/expired', [PasswordExpiredController::class, 'expired'])->name('password.expired');
+            Route::patch('password/expired', [PasswordExpiredController::class, 'update'])->name('password.expired.update');
         }
     });
 
@@ -33,28 +42,28 @@ Route::group(['namespace' => 'Auth', 'as' => 'auth.'], function () {
      */
     Route::group(['middleware' => 'guest'], function () {
         // Authentication Routes
-        Route::get('login', 'LoginController@showLoginForm')->name('login');
-        Route::post('login', 'LoginController@login')->name('login.post');
+        Route::get('login', [LoginController::class, 'showLoginForm'])->name('login');
+        Route::post('login', [LoginController::class, 'login'])->name('login.post');
 
         // Socialite Routes
-        Route::get('login/{provider}', 'SocialLoginController@login')->name('social.login');
-        Route::get('login/{provider}/callback', 'SocialLoginController@login');
+        Route::get('login/{provider}', [SocialLoginController::class, 'login'])->name('social.login');
+        Route::get('login/{provider}/callback', [SocialLoginController::class, 'login']);
 
         // Registration Routes
         if (config('access.registration')) {
-            Route::get('register', 'RegisterController@showRegistrationForm')->name('register');
-            Route::post('register', 'RegisterController@register')->name('register.post');
+            Route::get('register', [RegisterController::class, 'showRegistrationForm'])->name('register');
+            Route::post('register', [RegisterController::class, 'register'])->name('register.post');
         }
 
         // Confirm Account Routes
-        Route::get('account/confirm/{token}', 'ConfirmAccountController@confirm')->name('account.confirm');
-        Route::get('account/confirm/resend/{uuid}', 'ConfirmAccountController@sendConfirmationEmail')->name('account.confirm.resend');
+        Route::get('account/confirm/{token}', [ConfirmAccountController::class, 'confirm'])->name('account.confirm');
+        Route::get('account/confirm/resend/{uuid}', [ConfirmAccountController::class, 'sendConfirmationEmail'])->name('account.confirm.resend');
 
         // Password Reset Routes
-        Route::get('password/reset', 'ForgotPasswordController@showLinkRequestForm')->name('password.email');
-        Route::post('password/email', 'ForgotPasswordController@sendResetLinkEmail')->name('password.email.post');
+        Route::get('password/reset', [ForgotPasswordController::class, 'showLinkRequestForm'])->name('password.email');
+        Route::post('password/email', [ForgotPasswordController::class, 'sendResetLinkEmail'])->name('password.email.post');
 
-        Route::get('password/reset/{token}', 'ResetPasswordController@showResetForm')->name('password.reset.form');
-        Route::post('password/reset', 'ResetPasswordController@reset')->name('password.reset');
+        Route::get('password/reset/{token}', [ResetPasswordController::class, 'showResetForm'])->name('password.reset.form');
+        Route::post('password/reset', [ResetPasswordController::class, 'reset'])->name('password.reset');
     });
 });
