@@ -1,12 +1,18 @@
 <?php
 
-/**
+use App\Http\Controllers\Frontend\HomeController;
+use App\Http\Controllers\Frontend\ContactController;
+use App\Http\Controllers\Frontend\User\AccountController;
+use App\Http\Controllers\Frontend\User\ProfileController;
+use App\Http\Controllers\Frontend\User\DashboardController;
+
+/*
  * Frontend Controllers
  * All route names are prefixed with 'frontend.'.
  */
-Route::get('/', 'HomeController@index')->name('index');
-Route::get('contact', 'ContactController@index')->name('contact');
-Route::post('contact/send', 'ContactController@send')->name('contact.send');
+Route::get('/', [HomeController::class, 'index'])->name('index');
+Route::get('contact', [ContactController::class, 'index'])->name('contact');
+Route::post('contact/send', [ContactController::class, 'send'])->name('contact.send');
 
 /*
  * These frontend controllers require the user to be logged in
@@ -18,16 +24,16 @@ Route::group(['middleware' => ['auth', 'password_expires']], function () {
         /*
          * User Dashboard Specific
          */
-        Route::get('dashboard', 'DashboardController@index')->name('dashboard');
+        Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard');
 
         /*
          * User Account Specific
          */
-        Route::get('account', 'AccountController@index')->name('account');
+        Route::get('account', [AccountController::class, 'index'])->name('account');
 
         /*
          * User Profile Specific
          */
-        Route::patch('profile/update', 'ProfileController@update')->name('profile.update');
+        Route::patch('profile/update', [ProfileController::class, 'update'])->name('profile.update');
     });
 });
