@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; السابق',
-    'next'     => 'التالي &raquo;',
+    'previous' => '« السابق',
+    'next'     => 'التالي »',
 ];
