@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Previous',
-    'next'     => 'Next &raquo;',
+    'previous' => '« Previous',
+    'next'     => 'Next »',
 ];
