@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; قبلی',
-    'next'     => 'بعدی &raquo;',
+    'previous' => '« قبلی',
+    'next'     => 'بعدی »',
 ];
