@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Précédent',
-    'next'     => 'Suivant &raquo;',
+    'previous' => '« Précédent',
+    'next'     => 'Suivant »',
 ];
