@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Sebelumnya',
-    'next'     => 'Selanjutnya &raquo;',
+    'previous' => '« Sebelumnya',
+    'next'     => 'Selanjutnya »',
 ];
