@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Vorige',
-    'next'     => 'Volgende &raquo;',
+    'previous' => '« Vorige',
+    'next'     => 'Volgende »',
 ];
