@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Forrige',
-    'next'     => 'Neste &raquo;',
+    'previous' => '« Forrige',
+    'next'     => 'Neste »',
 ];
