@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'previous' => '&laquo; Назад',
-    'next'     => 'Вперёд &raquo;',
+    'previous' => '« Назад',
+    'next'     => 'Вперёд »',
 
 ];
