@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; ก่อนหน้า',
-    'next'     => 'ถัดไป &raquo;',
+    'previous' => '« ก่อนหน้า',
+    'next'     => 'ถัดไป »',
 ];
