@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; Önceki',
-    'next'     => 'Sonraki &raquo;',
+    'previous' => '« Önceki',
+    'next'     => 'Sonraki »',
 ];
