@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; 上一頁',
-    'next' => '下一頁 &raquo;',
+    'previous' => '« 上一頁',
+    'next' => '下一頁 »',
 ];
