@@ -13,6 +13,6 @@ return [
     |
     */
 
-    'previous' => '&laquo; 上一页',
-    'next'     => '下一页 &raquo;',
+    'previous' => '« 上一页',
+    'next'     => '下一页 »',
 ];
