@@ -1,12 +1,14 @@
 <?php
 
-/**
+use App\Http\Controllers\LanguageController;
+
+/*
  * Global Routes
  * Routes that are used between both frontend and backend.
  */
 
 // Switch between the included languages
-Route::get('lang/{lang}', 'LanguageController');
+Route::get('lang/{lang}', [LanguageController::class, 'swap']);
 
 /*
  * Frontend Routes
